refactor: ResultScreen を screen/container/layout の 4 層構成へ移行する
- usecase/result に rankFinalScores (順位付け + calcFinalScore による
  最終得点算出) を新設し Vitest 4 件を追加
- ResultScreen.tsx を screen (composition) / container (usecase 呼び出し
  と view model 変換) / layout (純 JSX) に分割し、結果 1 行の描画と
  表示ヘルパを component/RankingRow へ抽出
- .gitignore の result パターンを /result にアンカーし、
  src/usecase/result/ が誤って ignore されるのを修正
